Drop faulty + unused js code
The code gives an error (TypeError: select is null) as it needs a
if check to test if the value is saved. So the event handler is never
run. Despite this, the LAST_IDP_COOKIE setting puts in a cookie
with the last picked IDP so it keeps on working.